separate operations into mod
