+++ src/main.rs
@@ -1,12 +1,15 @@
 use anyhow::Error;
 use clap::Parser;
 use const_format::formatcp;
-use player::{Player, Serialisable};
-use std::collections::HashSet;
-use std::fs::{self, File};
-use std::io::Write;
-use std::ptr;
+use operations::{
+    add, delay, exit, load, pause, play, readline, remove, save, set_end, set_start, set_volume,
+    show, stop, toggle_loop, unloop, RespondResult,
+};
+use player::Player;
 use std::{path::PathBuf, time::Duration};
+
+mod operations;
+mod player;
 
 //TODO: propper error messages
 //TODO: reset command
@@ -200,8 +203,6 @@ fn parse_duration(dur: &str) -> Result<Duration, Error> {
     Ok(duration_str::parse(dur)?)
 }
 
-mod player;
-
 fn main() -> Result<(), String> {
     let mut players = Vec::new();
     let mut has_been_saved = true;
@@ -230,36 +231,6 @@ fn main() -> Result<(), String> {
     }
 }
 
-macro_rules! apply_vec {
-    ($vec:expr, $($inside:tt)*) => {
-        for x in $vec {
-            apply_vec!(x; $($inside)*);
-        }
-    };
-    ($x:ident; $func:ident($($arg:expr),*) $(,$($more:tt)*)?) => {
-        $x.$func($($arg),*);
-        $(apply_vec!($x; $($more)*))?
-    };
-    ($x:ident; $func:ident($($arg:expr),*) ? $(,$($more:tt)*)?) => {
-        $x.$func($($arg),*)?;
-        $(apply_vec!($x; $($more)*))?
-    };
-}
-
-macro_rules! show_selection {
-    ($selection:expr) => {
-        for player in $selection {
-            println!("{}", player.to_string())
-        }
-    };
-}
-
-struct RespondResult {
-    mutated: bool,
-    saved: bool,
-    quit: bool,
-}
-
 fn respond(
     players: &mut Vec<Player>,
     line: &str,
@@ -269,263 +240,21 @@ fn respond(
         Error::msg("error: cannot parse input. Perhaps you have erronous quotation(\"\")?")
     })?;
     let matches = Commands::try_parse_from(args)?;
-    let mut mutated = false;
-    let mut saved = false;
     match matches {
-        Commands::Add { path, name } => {
-            if &name.to_lowercase() == "all" {
-                return Err(Error::msg(
-                    "error: you cannot use the name 'all', because it is a keyword.",
-                ));
-            }
-            if name.parse::<u32>().is_ok() {
-                return Err(Error::msg(format!(
-                    "error: you cannot use the name '{name}', because it is a number."
-                )));
-            }
-            if players.iter_mut().filter(|p| p.name == name).count() > 0 {
-                return Err(Error::msg(format!(
-                    "error: you cannot use the name '{name}', because it is already used."
-                )));
-            }
-            let new_player = Player::new(path, name)?;
-            println!("{}", new_player.to_string());
-            players.push(new_player);
-            mutated = true;
-        }
-        Commands::Remove { ids } => {
-            // unsafe to have a mut and not mut ref at the same time
-            // make sure to print before remove
-            // this is an ugly hack
-            let mut selected_players =
-                select_players(unsafe { &*(players as *const Vec<Player>) }, ids)?
-                    .iter()
-                    .map(|p| p.name.as_str())
-                    .enumerate()
-                    .collect::<Vec<(usize, &str)>>();
-            if get_confirmation("Are you sure you want to remove these sounds?")? {
-                println!(
-                    "Removed {}",
-                    selected_players
-                        .iter()
-                        .map(|(_, p)| *p)
-                        .collect::<Vec<&str>>()
-                        .join(", ")
-                );
-                selected_players.sort_by_key(|p| p.0);
-                selected_players.reverse();
-                for (pos, _) in selected_players.clone() {
-                    players.remove(pos);
-                }
-                mutated = true;
-            }
-        }
-        Commands::Play { ids } => {
-            let mut player = select_players_mut(players, ids)?;
-            apply_vec!(&mut player, play()?);
-            show_selection!(player);
-        }
-        Commands::Stop { ids } => {
-            let mut player = select_players_mut(players, ids)?;
-            apply_vec!(&mut player, stop());
-            show_selection!(player);
-        }
-        Commands::Pause { ids } => {
-            let mut player = select_players_mut(players, ids)?;
-            apply_vec!(&mut player, pause());
-            show_selection!(player);
-        }
-        Commands::Volume { ids, volume } => {
-            let mut player = select_players_mut(players, ids)?;
-            apply_vec!(&mut player, volume(volume));
-            show_selection!(player);
-            mutated = true;
-        }
-        Commands::Show { ids } => {
-            let player = select_players(players, ids)?;
-            show_selection!(player);
-        }
-        Commands::Loop { ids, duration } => {
-            let mut player = select_players_mut(players, ids)?;
-            apply_vec!(
-                &mut player,
-                toggle_loop(true),
-                loop_length(duration),
-                apply_settings_in_place(false)?
-            );
-            show_selection!(player);
-            mutated = true;
-        }
-        Commands::Unloop { ids } => {
-            let mut player = select_players_mut(players, ids)?;
-            apply_vec!(
-                &mut player,
-                toggle_loop(false),
-                apply_settings_in_place(false)?
-            );
-            show_selection!(player);
-            mutated = true;
-        }
-        Commands::SetStart { ids, pos: duration } => {
-            let mut player = select_players_mut(players, ids)?;
-            apply_vec!(
-                &mut player,
-                skip_duration(duration),
-                apply_settings_in_place(false)?
-            );
-            show_selection!(player);
-            mutated = true;
-        }
-        Commands::SetEnd { ids, pos: duration } => {
-            let mut player = select_players_mut(players, ids)?;
-            apply_vec!(
-                &mut player,
-                take_duration(duration),
-                apply_settings_in_place(false)?
-            );
-            show_selection!(player);
-            mutated = true;
-        }
-        Commands::Delay { ids, duration } => {
-            let mut player = select_players_mut(players, ids)?;
-            apply_vec!(
-                &mut player,
-                set_delay(duration),
-                apply_settings_in_place(false)?
-            );
-            show_selection!(player);
-            mutated = true;
-        }
-        Commands::Save { path } => {
-            let serialisable: Vec<Serialisable> =
-                players.iter_mut().map(|p| p.to_serializable()).collect();
-            let json = serde_json::to_string(&serialisable)?;
-            fs::write(path, json)?;
-            saved = true;
-        }
-        Commands::Load { path } => {
-            let add_to_soundscape = players.is_empty()
-                || get_confirmation("Do you want to add this to you current soundscape?")?;
-            let perform_action = add_to_soundscape
-                || has_been_saved
-                || get_confirmation("Are you sure you want to exit without saving?")?;
-            if perform_action {
-                let json: Vec<Serialisable> = serde_json::from_reader(File::open(path)?)?;
-                let new_players = json
-                    .into_iter()
-                    .map(|p| Player::from_serializable(&p))
-                    .collect::<Result<Vec<Player>, Error>>()?;
-                if !add_to_soundscape {
-                    players.clear();
-                }
-                players.extend(new_players);
-                show_selection!(players)
-            }
-        }
-        Commands::Exit => {
-            let perform_action = has_been_saved
-                || get_confirmation("Are you sure you want to exit without saving?")?;
-            if perform_action {
-                return Ok(RespondResult {
-                    mutated,
-                    saved,
-                    quit: true,
-                });
-            }
-        }
+        Commands::Add { path, name } => add(players, path, name),
+        Commands::Remove { ids } => remove(players, ids),
+        Commands::Play { ids } => play(players, ids),
+        Commands::Stop { ids } => stop(players, ids),
+        Commands::Pause { ids } => pause(players, ids),
+        Commands::Volume { ids, volume } => set_volume(players, ids, volume),
+        Commands::Show { ids } => show(players, ids),
+        Commands::Loop { ids, duration } => toggle_loop(players, ids, duration),
+        Commands::Unloop { ids } => unloop(players, ids),
+        Commands::SetStart { ids, pos: duration } => set_start(players, ids, duration),
+        Commands::SetEnd { ids, pos: duration } => set_end(players, ids, duration),
+        Commands::Delay { ids, duration } => delay(players, ids, duration),
+        Commands::Save { path } => save(players, path),
+        Commands::Load { path } => load(players, path, has_been_saved),
+        Commands::Exit => exit(has_been_saved),
     }
-    Ok(RespondResult {
-        mutated,
-        saved,
-        quit: false,
-    })
-}
-
-fn select_players(players: &[Player], ids: Vec<IdOrName>) -> Result<Vec<&Player>, Error> {
-    if players.is_empty() {
-        return Err(Error::msg(
-            "error: there are addcurrently no players to select",
-        ));
-    }
-    if ids.contains(&IdOrName::All) {
-        if ids[0] == IdOrName::All && ids.len() == 1 {
-            return Ok(players.iter().collect());
-        }
-        return Err(Error::msg("'all' has to be the only id."));
-    }
-    if ids.is_empty() {
-        return Ok(vec![players.last().unwrap()]);
-    }
-
-    let ids_set: HashSet<usize> = ids
-        .into_iter()
-        .map(|id| match id {
-            IdOrName::Id(num) => Ok(num),
-            IdOrName::Name(name) => players
-                .iter()
-                .position(|p| p.name == name)
-                .ok_or_else(|| Error::msg(format!("error: no player found with the name {name}"))),
-            IdOrName::All => unreachable!(),
-        })
-        .collect::<Result<HashSet<usize>, Error>>()?;
-
-    Ok(ids_set.into_iter().map(|id| &players[id]).collect())
-}
-
-fn select_players_mut(
-    players: &mut [Player],
-    ids: Vec<IdOrName>,
-) -> Result<Vec<&mut Player>, Error> {
-    if players.is_empty() {
-        return Err(Error::msg(
-            "error: there are currently no players to select",
-        ));
-    }
-    if ids.contains(&IdOrName::All) {
-        if ids[0] == IdOrName::All && ids.len() == 1 {
-            return Ok(players.iter_mut().collect());
-        }
-        return Err(Error::msg("'all' has to be the only id."));
-    }
-    if ids.is_empty() {
-        return Ok(vec![players.last_mut().unwrap()]);
-    }
-
-    let ids_set: HashSet<usize> = ids
-        .into_iter()
-        .map(|id| match id {
-            IdOrName::Id(num) => Ok(num),
-            IdOrName::Name(name) => players
-                .iter_mut()
-                .position(|p| p.name == name)
-                .ok_or_else(|| Error::msg(format!("error: no player found with the name {name}"))),
-            IdOrName::All => unreachable!(),
-        })
-        .collect::<Result<HashSet<usize>, Error>>()?;
-
-    // because of the hashset, I know that the id's will be unique
-    // meaning that borrowing multiple items in players as mutable is safe
-    // as long as I only borrow that ones corresponding to the ids
-    Ok(ids_set
-        .into_iter()
-        .map(|id| unsafe { &mut *ptr::addr_of_mut!(players[id]) })
-        .collect())
-}
-
-fn readline(prompt: &str) -> Result<String, String> {
-    write!(std::io::stdout(), "{prompt}").map_err(|e| e.to_string())?;
-    std::io::stdout().flush().map_err(|e| e.to_string())?;
-    let mut buffer = String::new();
-    std::io::stdin()
-        .read_line(&mut buffer)
-        .map_err(|e| e.to_string())?;
-    Ok(buffer)
-}
-
-fn get_confirmation(prompt: &str) -> Result<bool, Error> {
-    Ok(readline(format!("{prompt} Y/N: ").as_str())
-        .map_err(Error::msg)?
-        .trim()
-        .to_lowercase()
-        == "y")
 }
